Programmatically config Jest for Travis
