+++ package.json
@@ -3,21 +3,6 @@
   "version": "1.0.0",
   "description": "A calorie tracker project for Mod4 BE at Turing School of Software and Design",
   "main": "app.js",
-  "jest": {
-    "clearMocks": true,
-    "collectCoverage": true,
-    "coverageDirectory": "coverage",
-    "coveragePathIgnorePatterns": [
-      "/node_modules"
-    ],
-    "setupFilesAfterEnv": [
-      "<rootDir>/config/jest.config"
-    ],
-    "testEnvironment": "node",
-    "testPathIgnorePatterns": [
-      "/node_modules"
-    ]
-  },
   "dependencies": {
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
@@ -36,7 +21,7 @@
   },
   "scripts": {
     "start": "node bin/www",
-    "test": "jest --runInBand",
+    "test": "jest --runInBand --config ./config/jest.config.js",
     "debug-test": "NODE_ENV=debug-test node --inspect-brk node_modules/.bin/jest --runInBand"
   },
   "repository": {
